requirements.txt
Flask
requests
PyYAML
uWSGI
